Code housecleaning; reworked module structure removed unused files from 1.0 implementation
