--- micro_dl/input/__init__.py
@@ -1,5 +1,0 @@
-"""Module for Input functions"""
-# from .dataset import BaseDataSet # To be replaced by gunpowder.
-# from .dataset_with_mask import DataSetWithMask
-# from .inference_dataset import InferenceDataSet
-from .training_table import BaseTrainingTable
