+++ micro_dl/input/dataset.py
@@ -1,509 +1,866 @@
-"""Dataset classes"""
-import warnings
+import glob
+from typing import Union
 import cv2
-# from tensorflow import keras
+import collections
+import copy
+import gunpowder as gp
 import numpy as np
 import os
-from scipy import ndimage
-import micro_dl.utils.normalize as norm
+import pandas as pd
+import torch
+from torch.utils.data import Dataset
+import zarr
+
+import micro_dl.utils.normalize as norm_utils
+import micro_dl.input.gunpowder_nodes as custom_nodes
+import micro_dl.utils.gunpowder_utils as gp_utils
+import micro_dl.utils.aux_utils as aux_utils
 
 
-def transform_matrix_offset_center(matrix, x, y):
-    o_x = float(x) / 2 - 0.5
-    o_y = float(y) / 2 - 0.5
-    offset_matrix = np.array([[1, 0, o_x], [0, 1, o_y], [0, 0, 1]])
-    reset_matrix = np.array([[1, 0, -o_x], [0, 1, -o_y], [0, 0, 1]])
-    transform_matrix = np.dot(np.dot(offset_matrix, matrix), reset_matrix)
-    return transform_matrix
-
-
-def apply_affine_transform(
-    x,
-    theta=0,
-    tx=0,
-    ty=0,
-    shear=0,
-    zx=1,
-    zy=1,
-    row_axis=1,
-    col_axis=2,
-    channel_axis=0,
-    fill_mode="nearest",
-    cval=0.0,
-    order=1,
-):
-    """Applies an affine transformation specified by the parameters given.
-    # Arguments
-        x: 3D numpy array - a 2D image with one or more channels.
-        theta: Rotation angle in degrees.
-        tx: Width shift.
-        ty: Heigh shift.
-        shear: Shear angle in degrees.
-        zx: Zoom in x direction.
-        zy: Zoom in y direction
-        row_axis: Index of axis for rows (aka Y axis) in the input image.
-                  Direction: left to right.
-        col_axis: Index of axis for columns (aka X axis) in the input image.
-                  Direction: top to bottom.
-        channel_axis: Index of axis for channels in the input image.
-        fill_mode: Points outside the boundaries of the input
-            are filled according to the given mode
-            (one of `{'constant', 'nearest', 'reflect', 'wrap'}`).
-        cval: Value used for points outside the boundaries
-            of the input if `mode='constant'`.
-        order: int, order of interpolation
-    # Returns
-        The transformed version of the input.
+class TorchDatasetContainer(object):
     """
-    # Input sanity checks:
-    # 1. x must 2D image with one or more channels (i.e., a 3D tensor)
-    # 2. channels must be either first or last dimension
-    if np.unique([row_axis, col_axis, channel_axis]).size != 3:
-        raise ValueError(
-            "'row_axis', 'col_axis', and 'channel_axis'" " must be distinct"
+    Dataset container object which initalizes multiple TorchDatasets depending upon the parameters
+    given in the training and network config files during initialization.
+
+    By default, randomly selects samples to divy up sources between train, test, and validation datasets.
+    """
+
+    def __init__(
+        self,
+        zarr_dir,
+        train_config,
+        network_config,
+        dataset_config,
+        device=None,
+        workers=0,
+        use_recorded_split=False,
+        data_split={},
+        allow_augmentations=True,
+    ):
+        """
+        Inits an object which builds a testing, training, and validation dataset
+        from .zarr data files using gunpowder pipelines
+
+        If acting as a container object:
+        :param dict train_config: dict object of train_config
+        :param dict network_config: dict object of network_config
+        :param dict dataset_config: dict object of dataset_config
+        :param str device: device on which to place tensors in child datasets,
+                            by default, places on 'cuda'
+        :param int workers: number of cpu workers for simultaneous data fetching
+        :param bool use_recorded_split: whether to use random or previously recorded
+                            train/test/val data split
+        :param dict data_split: dictionary of data split containing integer list of positions
+                            OR decimal fractions indicating split under {'train', 'test', 'val'}
+                            keys
+        :param bool allow_augmentations: If False disallows augmentations in dataloading
+                            pipelines, by default True
+        """
+        self.zarr_dir = zarr_dir
+        self.train_config = train_config
+        self.network_config = network_config
+        self.dataset_config = dataset_config
+        self.device = device
+        self.workers = workers
+        self.data_split = data_split
+
+        # acquire sources from the zarr directory
+        array_spec = gp_utils.generate_array_spec(network_config)
+        (
+            self.train_source,
+            self.test_source,
+            self.val_source,
+            self.dataset_keys,
+            self.data_split,
+        ) = gp_utils.multi_zarr_source(
+            zarr_dir=self.zarr_dir,
+            array_spec=array_spec,
+            data_split=self.data_split,
+            use_recorded_split=use_recorded_split,
         )
-
-    # TODO: shall we support negative indices?
-    valid_indices = set([0, 1, 2])
-    actual_indices = set([row_axis, col_axis, channel_axis])
-    if actual_indices != valid_indices:
-        raise ValueError(f"Invalid axis' indices: {actual_indices - valid_indices}")
-
-    if x.ndim != 3:
-        raise ValueError("Input arrays must be multi-channel 2D images.")
-    if channel_axis not in [0, 2]:
-        raise ValueError("Channels are allowed and the first and last dimensions.")
-
-    transform_matrix = None
-    if theta != 0:
-        theta = np.deg2rad(theta)
-        rotation_matrix = np.array(
-            [
-                [np.cos(theta), -np.sin(theta), 0],
-                [np.sin(theta), np.cos(theta), 0],
-                [0, 0, 1],
-            ]
-        )
-        transform_matrix = rotation_matrix
-
-    if tx != 0 or ty != 0:
-        shift_matrix = np.array([[1, 0, tx], [0, 1, ty], [0, 0, 1]])
-        if transform_matrix is None:
-            transform_matrix = shift_matrix
-        else:
-            transform_matrix = np.dot(transform_matrix, shift_matrix)
-
-    if shear != 0:
-        shear = np.deg2rad(shear)
-        shear_matrix = np.array(
-            [[1, -np.sin(shear), 0], [0, np.cos(shear), 0], [0, 0, 1]]
-        )
-        if transform_matrix is None:
-            transform_matrix = shear_matrix
-        else:
-            transform_matrix = np.dot(transform_matrix, shear_matrix)
-
-    if zx != 1 or zy != 1:
-        zoom_matrix = np.array([[zx, 0, 0], [0, zy, 0], [0, 0, 1]])
-        if transform_matrix is None:
-            transform_matrix = zoom_matrix
-        else:
-            transform_matrix = np.dot(transform_matrix, zoom_matrix)
-
-    if transform_matrix is not None:
-        h, w = x.shape[row_axis], x.shape[col_axis]
-        transform_matrix = transform_matrix_offset_center(transform_matrix, h, w)
-        x = np.rollaxis(x, channel_axis, 0)
-
-        # Matrix construction assumes that coordinates are x, y (in that order).
-        # However, regular numpy arrays use y,x (aka i,j) indexing.
-        # Possible solution is:
-        #   1. Swap the x and y axes.
-        #   2. Apply transform.
-        #   3. Swap the x and y axes again to restore image-like data ordering.
-        # Mathematically, it is equivalent to the following transformation:
-        # M' = PMP, where P is the permutation matrix, M is the original
-        # transformation matrix.
-        if col_axis > row_axis:
-            transform_matrix[:, [0, 1]] = transform_matrix[:, [1, 0]]
-            transform_matrix[[0, 1]] = transform_matrix[[1, 0]]
-        final_affine_matrix = transform_matrix[:2, :2]
-        final_offset = transform_matrix[:2, 2]
-
-        channel_images = [
-            ndimage.interpolation.affine_transform(
-                x_channel,
-                final_affine_matrix,
-                final_offset,
-                order=order,
-                mode=fill_mode,
-                cval=cval,
+        try:
+            assert len(self.test_source) > 0
+            assert len(self.train_source) > 0
+            assert len(self.val_source) > 0
+        except Exception as e:
+            raise AssertionError(
+                f"All datasets must have at least one source node / zarr store,"
+                f" not enough source arrays found.\n {e.args}"
             )
-            for x_channel in x
-        ]
-        x = np.stack(channel_images, axis=0)
-        x = np.rollaxis(x, 0, channel_axis + 1)
-    return x
+
+        # extract special keytypes
+        self.flatfield_key = None
+        for key in self.dataset_keys.copy():
+            if key.identifier == "flatfield":
+                self.flatfield_key = key
+                self.dataset_keys.remove(self.flatfield_key)
+                break
+
+        self.mask_key = None
+        mask_key_identifier = ""
+        if "mask_type" in dataset_config:
+            mask_key_identifier = "_".join(["mask", dataset_config["mask_type"]])
+
+        for key in self.dataset_keys.copy():
+            if key.identifier == mask_key_identifier:
+                self.mask_key = key
+                self.dataset_keys.remove(self.mask_key)
+            elif "mask" in key.identifier:
+                self.dataset_keys.remove(key)
+        if "mask_type" in dataset_config and self.mask_key == None:
+            raise ValueError(f"Specified mask type's corresponding array not found")
+
+        # build augmentation nodes if allowed
+        train_aug_nodes = []
+        test_aug_nodes = []
+        val_aug_nodes = []
+        if "augmentations" in train_config and allow_augmentations:
+            assert isinstance(train_config["augmentations"], dict), "Augmentations"
+            " section of config must be a dictionary of parameters"
+            train_aug_nodes = gp_utils.generate_augmentation_nodes(
+                train_config["augmentations"], self.dataset_keys
+            )
+            test_aug_nodes = gp_utils.generate_augmentation_nodes(
+                train_config["augmentations"], self.dataset_keys
+            )
+            val_aug_nodes = gp_utils.generate_augmentation_nodes(
+                train_config["augmentations"], self.dataset_keys
+            )
+
+        # set up epoch length for each dataset type or allow for later calculation
+        if "samples_per_epoch" in self.train_config:
+            train_epoch_length = self.train_config["samples_per_epoch"]
+        else:
+            train_epoch_length = 0
+        train_fraction = self.dataset_config["split_ratio"]["train"]
+        total = train_epoch_length // train_fraction
+
+        test_epoch_length = int(self.dataset_config["split_ratio"]["test"] * total)
+        val_epoch_length = int(self.dataset_config["split_ratio"]["val"] * total)
+
+        # assign each source subset to a child dataset object
+        self.train_dataset = self.init_torch_dataset(
+            self.train_source, train_aug_nodes, train_epoch_length
+        )
+        self.test_dataset = self.init_torch_dataset(
+            self.test_source, test_aug_nodes, test_epoch_length
+        )
+        self.val_dataset = self.init_torch_dataset(
+            self.val_source, val_aug_nodes, val_epoch_length
+        )
+
+    def init_torch_dataset(self, source, augmentation_nodes, dataset_epoch_length):
+        """
+        Initializes a torch dataset to sample 'source' tuple through the given
+        augmentations 'augmentations'.
+
+        :param tuple(gp.ZarrSource) source: tuple of source nodes representing the
+                                            dataset sample space
+        :param list augmentation_nodes: list of augmentation nodes in order
+        :param int dataset_epoch_length: length of epoch for this dataset.
+        """
+        # NOTE: not passing the whole dataset config is a design choice here. The
+        # elements of the config are a black box until theyre indexed. I do this
+        # to make them more readable. This can change with PyDantic later
+        dataset = TorchDataset(
+            # gunpowder params
+            data_source=source,
+            augmentation_nodes=augmentation_nodes,
+            data_keys=self.dataset_keys,
+            # preprocessing node params
+            mask_key=self.mask_key,
+            flatfield_key=self.flatfield_key,
+            normalization_scheme=self.dataset_config["normalization"]["scheme"],
+            normalization_type=self.dataset_config["normalization"]["type"],
+            min_foreground_fraction=self.dataset_config["min_foreground_fraction"],
+            flatfield_correct=self.dataset_config["flatfield_correct"],
+            # dataloading params
+            data_dimensionality=self.network_config["architecture"],
+            batch_size=self.dataset_config["batch_size"],
+            epoch_length=dataset_epoch_length,
+            target_channel_idx=tuple(self.dataset_config["target_channels"]),
+            input_channel_idx=tuple(self.dataset_config["input_channels"]),
+            spatial_window_size=tuple(self.dataset_config["window_size"]),
+            spatial_window_offset=(0,) * len(tuple(self.dataset_config["window_size"])),
+            device=self.device,
+            workers=self.workers,
+            random_sampling=True,
+        )
+        return dataset
+
+    def __getitem__(self, idx):
+        """
+        Provides indexing capabilities to reference train, test, and val datasets
+
+        :param int or str idx: index/key of dataset to retrieve:
+                                train -> 0 or 'train'
+                                test -> 1 or 'test'
+                                val -> 2 or 'val'
+        """
+        if isinstance(idx, str):
+            return {
+                "train": self.train_dataset,
+                "test": self.test_dataset,
+                "val": self.val_dataset,
+            }[idx]
+        else:
+            return [self.train_dataset, self.test_dataset, self.val_dataset][idx]
 
 
-# class BaseDataSet(keras.utils.Sequence):
-#     # warnings.warn('Warning: tf-dependent BaseDataSet to be replaced with GunPowder in 2.1.0')
-#     """
-#     Base class for input and target images
+class InferenceDatasetContainer(object):
+    """
+    Dataset container object which initalizes multiple TorchDatasets depending upon the
+    parametersgiven in the training, inference and network config files during initialization.
 
-#     https://www.tensorflow.org/api_docs/python/tf/keras/utils/Sequence
-#     https://stanford.edu/~shervine/blog/keras-how-to-generate-data-on-the-fly.html
-#     https://github.com/aleju/imgaug
-#     """
+    Unlike the TorchDatasetObject meant for training, this object does not initialize any
+    random sampling. This means that instead of initializing one dataset which randomly
+    accesses many positions, it returns a labeled list of datasets: one for each source
+    at each z-slice.
 
-#     def __init__(
-#         self,
-#         tile_dir,
-#         input_fnames,
-#         target_fnames,
-#         dataset_config,
-#         batch_size,
-#         image_format="zyx",
-#     ):
-#         """Init
+    Dataset objects are relatively light-weight until they are called, so having a large number
+    should not affect performance drastically.
 
-#         The images could be normalized at the image level during tiling
-#         (default, normalize=False). If images were not normalized during tiling
-#         set the normalize flag to normalize at the tile-level
-#         Works for tile shapes with z first (zyx) or last (xyz).
-#         Tiles will be loaded as is, and shape order will be determined
-#         from the preprocessing config file which is saved in your training
-#         data dir as preprocessing_info.json.
+    Requires use of a recorded data split to determine train, test, and validation positions
+    """
 
-#         :param str tile_dir: directory containing training image tiles
-#         :param pd.Series input_fnames: pd.Series with each row containing
-#          filenames for one input
-#         :param pd.Series target_fnames: pd.Series with each row containing
-#          filenames for one target
-#         :param dict dataset_config: Dataset part of the main config file
-#             Can contain a subset augmentations with args see line 186
-#         :param int batch_size: num of datasets in each batch
-#         :param str image_format: Tile shape order: 'xyz' or 'zyx'
-#         """
-#         self.tile_dir = tile_dir
-#         self.input_fnames = input_fnames
-#         self.target_fnames = target_fnames
-#         self.num_samples = len(self.input_fnames)
-#         self.batch_size = batch_size
-#         assert image_format in {
-#             "xyz",
-#             "zyx",
-#         }, "Image format should be xyz or zyx, not {}".format(image_format)
-#         self.image_format = image_format
+    def __init__(
+        self,
+        zarr_dir,
+        network_config,
+        dataset_config,
+        inference_config,
+        data_split,
+        device=None,
+        workers=0,
+    ):
+        """
+        Inits an object which rebuilds a previously recorded testing, training, and
+        validation dataset from .zarr data files using gunpowder pipelines
 
-#         # Check if model task (regression or segmentation) is specified
-#         self.model_task = "regression"
-#         if "model_task" in dataset_config:
-#             self.model_task = dataset_config["model_task"]
-#             assert self.model_task in {
-#                 "regression",
-#                 "segmentation",
-#             }, "Model task must be either 'segmentation' or 'regression'"
+        :param str zarr_dir: path to OME-Zarr HCS compatible zarr directory contraining
+                    training data
+        :param dict network_config: dict object of network_config
+        :param dict dataset_config: dict object of dataset_config
+        :param dict inference_config: dict object of inference_config
+        :param dict data_split: dictionary of data split containing integer list of positions
+                            OR decimal fractions indicating split under {'train', 'test', 'val'}
+                            keys
+        :param str device: device on which to place tensors in child datasets,
+                            by default, places on 'cuda'
+        :param int workers: number of cpu workers for simultaneous data fetching
+        """
+        self.zarr_dir = zarr_dir
+        self.network_config = network_config
+        self.dataset_config = dataset_config
+        self.inference_config = inference_config
+        self.data_split = data_split
+        self.device = device
+        self.workers = workers
 
-#         # Whether or not to do augmentations
-#         self.augmentations = False
-#         self.zoom_range = (1, 1)
-#         self.rotate_range = 0
-#         self.mean_jitter = 0
-#         self.std_jitter = 0
-#         self.noise_std = 0
-#         self.blur_range = (0, 0)
-#         self.shear_range = 0
-#         if "augmentations" in dataset_config:
-#             self.augmentations = True
-#             if "zoom_range" in dataset_config["augmentations"]:
-#                 self.zoom_range = dataset_config["augmentations"]["zoom_range"]
-#             if "rotate_range" in dataset_config["augmentations"]:
-#                 self.rotate_range = dataset_config["augmentations"]["rotate_range"]
-#             if "intensity_jitter" in dataset_config["augmentations"]:
-#                 self.mean_jitter, self.std_jitter = dataset_config["augmentations"][
-#                     "intensity_jitter"
-#                 ]
-#             if "noise_std" in dataset_config["augmentations"]:
-#                 self.noise_std = dataset_config["augmentations"]["noise_std"]
-#             if "blur_range" in dataset_config["augmentations"]:
-#                 self.blur_range = dataset_config["augmentations"]["blur_range"]
-#             if "shear_range" in dataset_config["augmentations"]:
-#                 self.shear_range = dataset_config["augmentations"]["shear_range"]
+        # acquire sources from the zarr directory
+        array_spec = gp_utils.generate_array_spec(network_config)
+        z_offset_range = self.calculate_z_offset_range()
+        (
+            self.train_sources,
+            self.test_sources,
+            self.val_sources,
+            self.dataset_keys,
+            self.data_split,
+        ) = gp_utils.multi_zarr_source(
+            zarr_dir=self.zarr_dir,
+            array_spec=array_spec,
+            use_recorded_split=True,
+            data_split=data_split,
+            copies=len(z_offset_range),
+        )
+        try:
+            assert len(self.test_sources) > 0
+            assert len(self.train_sources) > 0
+            assert len(self.val_sources) > 0
+        except Exception as e:
+            raise AssertionError(
+                f"All datasets must have at least one source node / zarr store,"
+                f" not enough source arrays found.\n {e.args}"
+            )
 
-#         # Whether to do zscore normalization on tile level
-#         self.normalize = False
-#         if "normalize" in dataset_config:
-#             self.normalize = dataset_config["normalize"]
-#         assert isinstance(self.normalize, bool), "normalize parameter should be boolean"
+        # extract special keytypes
+        self.flatfield_key = None
+        for key in self.dataset_keys.copy():
+            if key.identifier == "flatfield":
+                self.flatfield_key = key
+                self.dataset_keys.remove(self.flatfield_key)
+                break
 
-#         # Whether to shuffle indices at the end of each epoch
-#         self.shuffle = True
-#         if "shuffle" in dataset_config:
-#             assert isinstance(
-#                 dataset_config["shuffle"], bool
-#             ), "shuffle parameter should be boolean"
-#             self.shuffle = dataset_config["shuffle"]
+        self.mask_key = None
+        mask_key_identifier = ""
+        if "mask_type" in dataset_config:
+            mask_key_identifier = "_".join(["mask", dataset_config["mask_type"]])
 
-#         # Whether to only use a fraction of training data each epoch
-#         self.num_epoch_samples = self.num_samples
-#         if "train_fraction" in dataset_config:
-#             train_fraction = dataset_config["train_fraction"]
-#             assert (
-#                 0.0 < train_fraction <= 1.0
-#             ), "Train fraction should be [0,1], not {}".format(train_fraction)
-#             # You must shuffle if only using a fraction of the training data
-#             self.shuffle = True
-#             self.num_epoch_samples = int(self.num_samples * train_fraction)
-#         self.steps_per_epoch = int(
-#             np.ceil(
-#                 self.num_epoch_samples / self.batch_size,
-#             )
-#         )
-#         # Declare row indices, will to an inital shuffle at the end of init`
-#         self.row_idx = np.arange(self.num_samples)
+        for key in self.dataset_keys.copy():
+            if key.identifier == mask_key_identifier:
+                self.mask_key = key
+                self.dataset_keys.remove(self.mask_key)
+            elif "mask" in key.identifier:
+                self.dataset_keys.remove(key)
+        if "mask_type" in dataset_config and self.mask_key == None:
+            raise ValueError(f"Specified mask type's corresponding array not found")
 
-#         # Whether to remove singleton dimensions from tiles (e.g. 2D models)
-#         self.squeeze = False
-#         if "squeeze" in dataset_config:
-#             assert isinstance(
-#                 dataset_config["squeeze"], bool
-#             ), "squeeze parameter should be boolean"
-#             self.squeeze = dataset_config["squeeze"]
+        # assign each source subset to a child dataset object
+        self.train_dataset_list = self.generate_dataset_list(
+            self.train_sources,
+            z_offset_range,
+        )
+        self.test_dataset_list = self.generate_dataset_list(
+            self.test_sources,
+            z_offset_range,
+        )
+        self.val_dataset_list = self.generate_dataset_list(
+            self.val_sources,
+            z_offset_range,
+        )
 
-#         # Whether to use fixed random seed (only recommended for testing)
-#         self.random_seed = None
-#         if "random_seed" in dataset_config:
-#             self.random_seed = dataset_config["random_seed"]
-#         np.random.seed(self.random_seed)
+    def calculate_z_offset_range(self):
+        """
+        Gets the range of z-offsets for this data position to requested by the config setup.
 
-#         self.on_epoch_end()
+        :return list[range] z_offset_ranges: list of ranges containing z-slice offset ranges
+        """
+        z_slice_range = self.inference_config["z_slice_range"]
+        spatial_window_size = self.inference_config["window_size"]
+        spatial_window_z_depth = spatial_window_size[0]
+        z_offset_ranges = list(
+            range(z_slice_range[0], z_slice_range[1] - spatial_window_z_depth)
+        )
+        return z_offset_ranges
 
-#     def __len__(self):
-#         """Gets the number of batches per epoch"""
+    def generate_dataset_list(self, sources, z_offset_range):
+        """
+        Generate a list of datasets, one for each source at each z-slice in the specified
+        window range.
 
-#         n_batches = int(np.ceil(self.num_samples / self.batch_size))
-#         return n_batches
+        :param tuple(gp.ZarrSource) sources: multi-zarr-source (tuple of sources)
+        :param list[int] z_offset_range: z-offsets corresponding to each source
 
-#     def get_steps_per_epoch(self):
-#         """
-#         Returns steps per epoch which is number of training samples per
-#         epoch divided by batch size.
+        :return list datasets: returns a list of datasets each servicing one z-offset window
+                        in one source position
+        """
+        spatial_window_size = self.inference_config["window_size"]
 
-#         :return int steps_per_epoch: Steps per epoch
-#         """
-#         return self.steps_per_epoch
+        datasets = []
+        for source in sources:
+            for i, z_offset in enumerate(z_offset_range):
+                source_copy = source[i]
+                spatial_window_offset = (z_offset,) + (0,) * len(
+                    spatial_window_size[1:]
+                )
+                dataset = self.init_inference_dataset(
+                    (source_copy,),
+                    spatial_window_size,
+                    spatial_window_offset,
+                )
+                datasets.append(dataset)
 
-#     def _augment_image(
-#         self,
-#         input_image,
-#         aug_idx,
-#         zoom=1,
-#         theta=0,
-#         mean_offset=0,
-#         std_scale=1,
-#         noise_std=0,
-#         blur_img=False,
-#         blur_sigma=0,
-#         shear=0,
-#     ):
-#         """Adds image augmentation among 6 possible options
+        return datasets
 
-#         :param np.array input_image: input image to be transformed
-#         :param int aug_idx: integer specifying the transformation to apply.
-#          0 - Image as is
-#          1 - flip LR (horizontally) about axis 1 (y)
-#          2 - flip UD (vertically) about axis 0 (x)
-#          3 - rotate 90 degrees in the xy-plane in the x toward y direction
-#          4 - rotate 180 degrees in the xy-plane in the x toward y direction
-#          5 - rotate 270 degrees in the xy-plane in the x toward y direction
-#         :return np.array image after transformation is applied
-#         """
-#         if aug_idx == 0:
-#             return input_image
-#         elif aug_idx == 1:
-#             # flip about y (if zxy) or x if (zyx))
-#             trans_image = np.flip(input_image, -1)
-#         elif aug_idx == 2:
-#             # flip about x (if zxy) or y if (zyx))
-#             trans_image = np.flip(input_image, -2)
-#         elif aug_idx == 3:
-#             # rot in plane defined by axis=(-2,-1)
-#             trans_image = np.rot90(
-#                 input_image,
-#                 k=1,
-#                 axes=(-2, -1),
-#             )
-#         elif aug_idx == 4:
-#             # rot in plane defined by axis=(-2,-1)
-#             trans_image = np.rot90(
-#                 input_image,
-#                 k=2,
-#                 axes=(-2, -1),
-#             )
-#         elif aug_idx == 5:
-#             # rot in plane defined by axis=(-2,-1)
-#             trans_image = np.rot90(
-#                 input_image,
-#                 k=3,
-#                 axes=(-2, -1),
-#             )
-#         else:
-#             msg = "{} not in allowed aug_idx: 0-5".format(aug_idx)
-#             raise ValueError(msg)
-#         if blur_img:
-#             trans_image = cv2.GaussianBlur(trans_image, ksize=(0, 0), sigmaX=blur_sigma)
-#         if noise_std != 0:
-#             trans_image = trans_image + np.random.normal(
-#                 scale=noise_std, size=trans_image.shape
-#             )
-#         if not (mean_offset == 0 and std_scale == 1):
-#             trans_image = norm.unzscore(trans_image, mean_offset, std_scale)
-#         trans_image = apply_affine_transform(
-#             trans_image,
-#             zx=zoom,
-#             theta=theta,
-#             zy=zoom,
-#             shear=shear,
-#             fill_mode="constant",
-#             cval=0.0,
-#             order=1,
-#         )
-#         return trans_image
+    def init_inference_dataset(self, source, window_size, window_offset):
+        """
+        Initializes a torch dataset to sample 'source' tuple.
 
-#     def _get_volume(
-#         self,
-#         fname_list,
-#         normalize=True,
-#         aug_idx=0,
-#         zoom=1,
-#         theta=0,
-#         mean_offset=0,
-#         std_scale=1,
-#         noise_std=0,
-#         blur_img=False,
-#         blur_sigma=0,
-#         shear=0,
-#     ):
-#         """
-#         Read tiles from fname_list and stack them into an image volume.
+        :param tuple(gp.ZarrSource) source: tuple of source nodes representing the
+                                            dataset sample space
+        :param tuple(int) window_size: dataset sample size
+        :param tuple(int) window_offset: dataset sample window offset in data array
+        """
+        # NOTE: not passing the whole dataset config is a design choice here. The
+        # elements of the config are a black box until theyre indexed. I do this
+        # to make them more readable. This can change with PyDantic later
+        dataset = TorchDataset(
+            # gunpowder params
+            data_source=source,
+            augmentation_nodes=[],
+            data_keys=self.dataset_keys,
+            # preprocessing node params
+            mask_key=self.mask_key,
+            flatfield_key=self.flatfield_key,
+            normalization_scheme=self.dataset_config["normalization"]["scheme"],
+            normalization_type=self.dataset_config["normalization"]["type"],
+            min_foreground_fraction=0,
+            flatfield_correct=self.dataset_config["flatfield_correct"],
+            # dataloading params
+            data_dimensionality=self.network_config["architecture"],
+            batch_size=1,
+            epoch_length=1,
+            target_channel_idx=tuple(self.dataset_config["target_channels"]),
+            input_channel_idx=tuple(self.dataset_config["input_channels"]),
+            spatial_window_size=window_size,
+            spatial_window_offset=window_offset,
+            device=self.device,
+            workers=self.workers,
+            random_sampling=False,
+        )
+        return dataset
 
-#         :param list fname_list: list of file names of input/target images
-#         :param bool normalize: Whether to zscore normalize tiles
-#         :param int aug_idx: type of augmentation to be applied (if any)
-#         :return: np.ndarray of stacked images
-#         """
-#         image_volume = []
-#         for fname in fname_list:
-#             cur_tile = np.load(os.path.join(self.tile_dir, fname))
-#             if self.augmentations:
-#                 cur_tile = self._augment_image(
-#                     cur_tile,
-#                     aug_idx,
-#                     zoom=zoom,
-#                     theta=theta,
-#                     mean_offset=mean_offset,
-#                     std_scale=std_scale,
-#                     noise_std=noise_std,
-#                     blur_img=blur_img,
-#                     blur_sigma=blur_sigma,
-#                     shear=shear,
-#                 )
-#             if self.squeeze:
-#                 cur_tile = np.squeeze(cur_tile)
-#             image_volume.append(cur_tile)
-#         # Stack images channels first
-#         image_volume = np.stack(image_volume)
-#         if image_volume.dtype == bool:
-#             image_volume = image_volume.astype(np.float32)
-#         elif normalize:
-#             image_volume = norm.zscore(image_volume)
-#         return image_volume
+    def __getitem__(self, idx):
+        """
+        Provides indexing capabilities to reference train, test, and val datasets
 
-#     def __getitem__(self, index):
-#         """
-#         Get a batch of data. If using a fraction of the training data, shuffle
-#         if automatically set to True to make sure you have a chance of accessing
-#         all training data and not just the first indices.
+        :param int or str idx: index/key of dataset to retrieve:
+                                train -> 0 or 'train'
+                                test -> 1 or 'test'
+                                val -> 2 or 'val'
+        """
+        if isinstance(idx, str):
+            return {
+                "train": self.train_dataset_list,
+                "test": self.test_dataset_list,
+                "val": self.val_dataset_list,
+            }[idx]
+        else:
+            return [
+                self.train_dataset_list,
+                self.test_dataset_list,
+                self.val_dataset_list,
+            ][idx]
 
-#         https://www.tensorflow.org/performance/performance_guide#use_nchw_imag.
-#         will use nchw format. shape: [batch_size, num_channels, z, y, x]
 
-#         :param int index: batch index
-#         :return: np.ndarrays input_image and target_image of shape
-#          [batch_size, num_channels, z, y, x]
-#         """
+class TorchDataset(Dataset):
+    """
+    Based off of torch.utils.data.Dataset:
+        - https://pytorch.org/docs/stable/data.html
 
-#         start_idx = index * self.batch_size
-#         end_idx = (index + 1) * self.batch_size
-#         if end_idx >= self.num_samples:
-#             end_idx = self.num_samples
+    Custom dataset class that builds gunpowder pipelines composed of multi-zarr source nodes
+    and a series of augmentation nodes. This object will call from the gunpowder pipeline directly,
+    and transform resulting data into tensors to be placed onto the gpu for processing.
 
-#         norm_output = self.model_task != "segmentation" and self.normalize
+    Multiprocessing is supported with num_workers > 0. However, there are non-fatal warnings about
+    "...processes being terminated before shared CUDA tensors are released..." with torch 1.10.
 
-#         input_image = []
-#         target_image = []
-#         aug_idx = 0
-#         zoom = 1
-#         theta = 0
-#         mean_offset = 0
-#         std_scale = 1
-#         noise_std = 0
-#         blur_img = False
-#         blur_sigma = 0
-#         shear = 0
-#         for idx in range(start_idx, end_idx, 1):
-#             cur_input_fnames = self.input_fnames.iloc[self.row_idx[idx]]
-#             cur_target_fnames = self.target_fnames.iloc[self.row_idx[idx]]
-#             # Select select int randomly that will represent augmentation type
-#             if self.augmentations:
-#                 aug_idx = np.random.choice([0, 1, 2, 3, 4, 5], 1)
-#                 zoom = np.random.uniform(self.zoom_range[0], self.zoom_range[1])
-#                 theta = np.random.uniform(-self.rotate_range, self.rotate_range)
-#                 mean_offset = np.random.uniform(-self.mean_jitter, self.mean_jitter)
-#                 std_scale = 1 + np.random.uniform(-self.std_jitter, self.std_jitter)
-#                 noise_std = np.random.uniform(0, self.noise_std)
-#                 shear = np.random.uniform(-self.shear_range, self.shear_range)
-#                 if not (self.blur_range[0] == 0 and self.blur_range[1] == 0):
-#                     blur_img = np.random.choice([True, False], 1)[0]
-#                     blur_sigma = np.random.uniform(
-#                         self.blur_range[0], self.blur_range[1]
-#                     )
-#             # only apply intensity jitter to input
-#             cur_input = self._get_volume(
-#                 fname_list=cur_input_fnames.split(","),
-#                 normalize=self.normalize,
-#                 aug_idx=aug_idx,
-#                 zoom=zoom,
-#                 theta=theta,
-#                 mean_offset=mean_offset,
-#                 std_scale=std_scale,
-#                 noise_std=noise_std,
-#                 blur_img=blur_img,
-#                 blur_sigma=blur_sigma,
-#                 shear=shear,
-#             )
-#             cur_target = self._get_volume(
-#                 fname_list=cur_target_fnames.split(","),
-#                 normalize=norm_output,
-#                 aug_idx=aug_idx,
-#                 zoom=zoom,
-#                 theta=theta,
-#                 shear=shear,
-#                 mean_offset=0,
-#                 std_scale=1,
-#                 noise_std=0,
-#                 blur_img=False,
-#                 blur_sigma=0,
-#             )
-#             input_image.append(cur_input)
-#             target_image.append(cur_target)
+    These warnings are discussed on the following post, and I believe have been since fixed:
+        - https://github.com/pytorch/pytorch/issues/71187
+    """
 
-#         input_image = np.stack(input_image)
-#         target_image = np.stack(target_image)
-#         return input_image, target_image
+    def __init__(
+        self,
+        data_source,
+        augmentation_nodes,
+        data_keys,
+        mask_key,
+        flatfield_key,
+        normalization_scheme,
+        normalization_type,
+        min_foreground_fraction,
+        flatfield_correct,
+        data_dimensionality,
+        batch_size,
+        epoch_length,
+        target_channel_idx,
+        input_channel_idx,
+        spatial_window_size,
+        spatial_window_offset,
+        device,
+        workers,
+        random_sampling,
+    ):
+        """
+        Creates a dataset object which draws samples directly from a gunpowder pipeline.
 
-#     def on_epoch_end(self):
-#         """Update indices and shuffle after each epoch"""
-#         if self.shuffle:
-#             np.random.shuffle(self.row_idx)
+        :param tuple(gp.ZarrSource) source: tuple of source nodes representing the
+                                            dataset sample space
+        :param list augmentation_nodes: list of augmentation nodes in order
+        :param tuple data_source: tuple of gp.Source nodes which the given pipeline draws samples
+        :param list data_keys: list of gp.ArrayKey objects which access the given source
+        :param gp.ArrayKey mask_key: key to untracked mask array in source node, if applicable
+        :param gp.ArrayKey flatfield_key: key to untracked flatfield array in source note, if
+                                    applicable
+        :param str normalization_scheme: see name, one of {"dataset", "FOV", "tile"}
+        :param str normalization_type: see name, one of {"median_and_iqr", "mean_and_std"}
+        :param float min_foreground_fraction: minimum foreground required to be present in sample
+                                    for region to be selected, must be within [0, 1]
+        :param bool flatfield_correct: whether or not to flatfield correct
+        :param str data_dimensionality: whether to collapse the first channel of 3d data,
+                                    one of {2D, 2.5D}
+        :param int batch_size: number of samples per batch
+        :param tuple(int) target_channel_idx: indices of target channel(s) within zarr store
+        :param tuple(int) input_channel_idx: indices of input channel(s) within zarr store
+        :param tuple spatial_window_size: tuple of sample dimensions, specifies batch request ROI
+                                    expects 3D tuple; dimensions zyx, where z = 1 for 2d request
+        :param tuple spatial_window_offset: tuple of offset dimensions, specifies batch request ROI
+                                    expects 3D tuple; dimensions zyx, where z = 1 for 2d request
+        :param str device: device on which to place tensors in child datasets,
+                                    by default, places on 'cuda'
+        :param int workers: number of simultaneous threads reading data into batch requests
+        :param bool random_sampling: whether to sample source data randomly or not.
+        """
+        self.data_source = data_source
+        self.augmentation_nodes = augmentation_nodes
+        self.data_keys = data_keys
+        self.mask_key = mask_key
+        self.flatfield_key = flatfield_key
+        self.normalization_scheme = normalization_scheme
+        self.normalization_type = normalization_type
+        self.min_foreground_fraction = min_foreground_fraction
+        self.flatfield_correct = flatfield_correct
+        self.data_dimensionality = data_dimensionality
+        self.batch_size = batch_size
+        self.epoch_length = epoch_length
+        self.target_idx = target_channel_idx
+        self.input_idx = input_channel_idx
+        self.window_size = spatial_window_size
+        self.window_offset = spatial_window_offset
+        self.device = device
+        self.active_key = 0
+        self.workers = max(1, workers - 1)
+        self.random_sampling = random_sampling
+
+        # safety checks: iterate through keys and data sources to ensure that they match
+        voxel_size = None
+        for key in self.data_keys:
+            # check that all data voxel sizes are the same, exclude masks and flatfields
+            if not ("flatfield" in key.identifier or "mask" in key.identifier):
+                for i, source in enumerate(self.data_source):
+                    try:
+                        array_spec = source.array_specs[key]
+                        if not voxel_size:
+                            voxel_size = array_spec.voxel_size
+                        else:
+                            assert (
+                                voxel_size == array_spec.voxel_size
+                            ), f"Voxel size of array {array_spec.voxel_size} does not match"
+                            f" voxel size of previous array {voxel_size}."
+                    except Exception as e:
+                        raise AssertionError(
+                            f"Error matching keys to source in dataset: {e.args}"
+                        )
+
+        # calculate epoch length: if no epoch length specified, set according to data size
+        self._calculate_epoch_length()
+
+        # construct pipeline: generate batch request, construct nodes, make iterable
+        assert len(self.window_size) == 3, (
+            f"Window size {self.window_size} must be 3-dimensional. If 2D data, "
+            "desired, make the first dimension of window size 1; (1, X, Y)."
+        )
+        assert len(self.window_size) == len(voxel_size), (
+            f"Incompatible voxel size {voxel_size}. "
+            f"Must be same length as spatial_window_size {self.window_size}."
+        )
+        assert len(self.window_size) == len(self.window_offset), (
+            f"Incompatible window offset {self.window_offset}. "
+            f"Must be same length as spatial_window_size {self.window_size}."
+        )
+        self.batch_request = self._generate_batch_request(
+            window_offset=self.window_offset,
+            window_size=self.window_size,
+        )
+        self.preprocessing_nodes = self._generate_preprocessing_nodes()
+        self.pipeline = self._build_pipeline()
+
+        self.batch_generator = self._generate_batches()
+
+    def _calculate_epoch_length(self, oversample_factor=2):
+        """
+        Calculates the epoch length that should be used for this dataset based upon the
+        tile window size and the size of the data. Assumes that sampling will be uniformly
+        random over the spatial data, and that we will oversample our data to the specified
+        factor (default of 2).
+
+        :param int oversample_factor: factor with which we will oversample our data to ensure
+                            that data is given a fair chance at being at the "center" of a tile
+        """
+        if self.epoch_length == 0 or self.epoch_length == None:
+            source = self.data_source[0]
+            for key in source.datasets:
+                z1 = zarr.open(source.filename)
+                fov_shape = z1[source.datasets[key]][:].shape
+                break
+            data_spatial_dims = fov_shape[-len(self.window_size) :]
+
+            # assuming stride length providing oversampling, calculate # theoretical tiles/fov
+            samples_per_fov = 1
+            for dim_index in range(len(self.window_size)):
+                data_length = data_spatial_dims[dim_index]
+                if len(self.window_size) == 3 and dim_index == 0:  # if z
+                    samples_per_fov *= data_length - (self.window_size[dim_index] - 1)
+                else:  # if x or y
+                    samples_per_fov *= (
+                        (data_length // self.window_size[dim_index]) * oversample_factor
+                    ) - 1
+
+            self.epoch_length = int(samples_per_fov * len(self.data_source))
+            self.epoch_length = self.epoch_length // self.batch_size
+
+    def _generate_batch_request(self, window_offset, window_size):
+        """
+        Generates a series of batch requests according to a window offset and window size
+
+        :param tuple(int, int, int) window_offset: offset at start of window
+        :param tuple(int, int, int) window_size: size of dimensions of window
+
+        :return gp.BatchRequest() batch_request: batch request corresponding to given window
+        """
+        batch_request = gp.BatchRequest()
+
+        for key in self.data_keys:
+            batch_request[key] = gp.Roi(window_offset, window_size)
+            # NOTE: the keymapping we are performing here makes it so that if
+            # we DO end up generating multiple arrays at the position/well level
+            # (for example one with and without flatfield correction), we can
+            # access all of them by requesting that key. The index we request
+            # in __getitem__ ends up being the index of our key.
+        if self.flatfield_key:
+            batch_request[self.flatfield_key] = gp.Roi(
+                (0,) * len(window_offset), tuple([1] + list(window_size[1:]))
+            )
+        if self.mask_key:
+            batch_request[self.mask_key] = gp.Roi(window_offset, window_size)
+
+        return batch_request
+
+    def __len__(self):
+        """
+        Returns number of source data arrays in this dataset.
+        """
+        return self.epoch_length
+
+    def __getitem__(self, idx=0):
+        """
+        Requests a batch from the data pipeline using the key selected by self.use_key,
+        and applying that key to call a batch from its corresponding source using
+        self.batch_request.
+        """
+        assert self.active_key != None, "No active key. Try '.use_key()'"
+
+        # TODO: this implementation pulls 5 x 256 x 256 of all channels. We may not want
+        # all of those slices in all channels if theyre note being used. Fix this inefficiency
+
+        sample = next(self.batch_generator)
+        sample_data = sample[self.active_key].data
+
+        # NOTE We assume the .zarr ALWAYS has an extra batch channel.
+        # SO, 3d -> 5d data, 2d -> 4d data
+
+        # remove extra dimension from stack node, if 2d remove z dimension
+        sample_data = sample_data[:, 0, ...]
+        if self.window_size[0] == 1 and self.data_dimensionality == "2D":
+            sample_data = sample_data[..., 0, :, :]
+
+        # stack multiple channels
+        full_input = []
+        for idx in self.input_idx:  # assumes bczyx or bcyx
+            channel_input = sample_data[:, idx, ...]
+            full_input.append(channel_input)
+        full_input = np.stack(full_input, 1)
+
+        full_target = []
+        for idx in self.target_idx:  # assumes bczyx or bcyx
+            channel_target = sample_data[:, idx, ...]
+            full_target.append(channel_target)
+        full_target = np.stack(full_target, 1)
+
+        if len(full_target.shape) == 5:
+            # target is always 2 dimensional, we select middle z dim
+            middle_z_idx = full_target.shape[-3] // 2
+            full_target = np.expand_dims(full_target[..., middle_z_idx, :, :], -3)
+
+        # convert to tensor and place onto gpu
+        convert = ToTensor(self.device)
+        input_, target_ = convert(full_input), convert(full_target)
+
+        return (input_, target_)
+
+    def use_key(self, selection):
+        """
+        Sets self.active_key to selection if selection is an ArrayKey. If selection is an int,
+        sets self.active_key to the index in self.data_keys given by selection.
+
+        :param int or gp.ArrayKey selection: key index of key in self.data_keys to activate
+        """
+        if isinstance(selection, int):
+            try:
+                self.active_key = self.data_keys[selection]
+            except IndexError as e:
+                raise IndexError(
+                    f"Handling exception {e.args}: index of selection"
+                    " must be within length of data_keys"
+                )
+        elif isinstance(selection, gp.ArrayKey):
+            # TODO change this assertion to reference the source-> arrayspec-> keys
+            assert (
+                selection in self.data_keys
+            ), "Given key not associated with dataset sources"
+            self.active_key = selection
+        else:
+            raise AttributeError("Selection must be int or gp.ArrayKey")
+
+    def _build_pipeline(self):
+        """
+        Builds a gunpowder data pipeline given a source node and a list of augmentation
+        nodes.
+
+        :param gp.ZarrSource or tuple source: source node/tuple of source nodes from which to draw
+        :param list nodes: list of augmentation nodes, by default empty list
+
+        :return gp.Pipeline pipeline: see name
+        """
+        # ---- Sourcing Nodes ----#
+        # if source is multi_zarr_source, attach a RandomProvider
+        if self.random_sampling:
+            source = self.data_source
+            if isinstance(source, tuple):
+                source = [source, gp.RandomProvider()]
+            source = source + [gp.RandomLocation()]
+        else:
+            source = [self.data_source[0]]
+        # NOTE: not random sampling with multiple providers only calls from the first
+
+        if self.min_foreground_fraction and self.mask_key:
+            source = source + [
+                gp.Reject(
+                    mask=self.mask_key,
+                    min_masked=self.min_foreground_fraction,
+                ),
+                custom_nodes.PrepMaskRoi(self.data_keys, self.mask_key),
+            ]
+
+        # ---- Preprocessing Nodes ----#
+        source = source + self.preprocessing_nodes
+
+        # ---- Batch Creation Nodes ----#
+        batch_creation = []
+        # batch_creation.append(
+        #     gp.PreCache(
+        #         cache_size=self.epoch_length // 10,
+        #         num_workers=max(1, self.workers),
+        #     )
+        # )
+        batch_creation.append(gp.Stack(self.batch_size))
+
+        # attach additional nodes, if any, and sum
+        pipeline = source + self.augmentation_nodes + batch_creation
+        pipeline = gp_utils.gpsum(pipeline, verbose=False)
+        return pipeline
+
+    def _generate_batches(self):
+        """
+        Returns pipeline as a generator. This is done in a separate method from __getitem__()
+        to preserve compatibility with the torch dataloader's item calling signature
+        while also performing appropriate context management for the pipeline via generation.
+        See:
+            https://github.com/funkey/gunpowder/issues/181
+
+        :param gp.Pipeline pipeline: pipeline to generate batches from
+        :param gp.BatchRequest request: batch request for pipeline
+
+        :yield gp.Batch batch: single batch yielded from pipeline at each generation
+        :rtype: Iterator[gp.Batch]
+        """
+        with gp.build(self.pipeline):
+            while True:
+                yield self.pipeline.request_batch(self.batch_request)
+                # batch = self.pipeline.request_batch(self.batch_request)
+                # self.pipeline.internal_teardown()
+                # yield batch
+
+    def _generate_preprocessing_nodes(self):
+        """
+        Returns a list of preprocessing nodes configured according to the current state of
+        this object.
+
+        :return list preprocessing_nodes: see name
+        """
+        preprocessing_nodes = []
+
+        if self.flatfield_key and self.flatfield_correct:
+            flatfield_correct = custom_nodes.FlatFieldCorrect(
+                array=self.data_keys, flatfield=self.flatfield_key
+            )
+            preprocessing_nodes.append(flatfield_correct)
+
+        if self.normalization_scheme and self.normalization_type:
+            normalize = custom_nodes.Normalize(
+                array=self.data_keys,
+                scheme=self.normalization_scheme,
+                type=self.normalization_type,
+            )
+            preprocessing_nodes.append(normalize)
+
+        return preprocessing_nodes
+
+
+class DatasetEnsemble(TorchDataset):
+    def __init__(self, torch_datasets):
+        """
+        Groups the datasets in torch_datasets together, pulling from each of them
+        individually.
+
+        Used instead of ConcatDataset for context-management issues with gunpowder
+        generator pipeline
+
+        :param list torch_datasets: list of TorchDataset objects to sample from
+        :return: object instance
+        """
+        self.datasets = torch_datasets
+        self.last_pipeline = None
+
+    def __getitem__(self, idx):
+        """
+        Returns one sample by calling the __getitem__ of the subsequent TorchDataset at
+        index idx.
+
+        :param int idx: index of dataset to call sample from in torch_datasets
+        :return: item from __getitem__ of torch_datasets[idx]
+        """
+        dataset = self.datasets[idx]
+        self._check_source_uniqueness(dataset)
+
+        return dataset.__getitem__()
+
+    def _check_source_uniqueness(self, dataset):
+        dataset_pipeline = dataset.pipeline
+        if not self.last_pipeline:
+            self.last_pipeline = dataset_pipeline
+        else:
+            pipeline_1 = self.last_pipeline
+            pipeline_2 = dataset_pipeline
+
+            assert (
+                pipeline_1.output is not pipeline_2.output
+            ), f"Two nodes shared: {pipeline_1.output} is {pipeline_2.output}"
+
+            while len(pipeline_1.children) > 0:
+                pipeline_1 = pipeline_1.children[0]
+                pipeline_2 = pipeline_2.children[0]
+
+                assert (
+                    pipeline_1.output is not pipeline_2.output
+                ), f"Two nodes shared: {pipeline_1.output} is {pipeline_2.output}"
+
+            self.last_pipeline = dataset_pipeline
+
+    def __len__(self):
+        return len(self.datasets)
+
+
+class ToTensor(object):
+    """
+    Transformation. Converts input to torch.Tensor and returns. By default also places tensor
+    on gpu.
+
+    :param torch.device device: device transport tensor to
+    """
+
+    def __init__(self, device=torch.device("cuda")):
+        self.device = device
+
+    def __call__(self, sample):
+        """
+        Perform transformation.
+
+        :param torch.tensor or numpy.ndarray sample: data to convert to tensor and place on device
+        :return torch.tensor sample: converted data on device
+        """
+        if isinstance(sample, torch.Tensor):
+            sample = sample.to(self.device)
+        else:
+            sample = torch.tensor(sample, dtype=torch.float32).to(self.device)
+        return sample
