+++ micro_dl/preprocessing/generate_masks.py
@@ -7,7 +7,7 @@ import micro_dl.utils.image_utils as im_utils
 import micro_dl.utils.io_utils as io_utils
 from micro_dl.utils.mp_utils import mp_create_and_write_mask
 from skimage.filters import threshold_otsu
-from micro_dl.torch_unet.utils.io import MultiProcessProgressBar
+from micro_dl.utils.cli_utils import MultiProcessProgressBar
 
 
 class MaskProcessor:
