+++ micro_dl/utils/meta_utils.py
@@ -10,7 +10,7 @@ import micro_dl.utils.aux_utils as aux_utils
 import micro_dl.utils.image_utils as im_utils
 import micro_dl.utils.io_utils as io_utils
 import micro_dl.utils.mp_utils as mp_utils
-from micro_dl.torch_unet.utils.io import show_progress_bar
+from micro_dl.utils.cli_utils import show_progress_bar
 
 
 def generate_normalization_metadata(
